FEATURE: Add volume control to jack output
This adds a simple logarithmic soft volume control to the jack
output plugin. Config values are used to store the volume levels,
because jack provides no volume storage capability. Volume is controlled
in the usual way however and the config values need not be altered directly.
Zero crossings are not observed and amplitude changes can happen at any
time.