FEATURE: pitch effect plugin
There is a plugin called vocoder that does pitch things. It doesn't work
to well for me. This pitch plugin is based on secret rabbit code and thus
is a GPL plugin. It supports S16/S32/FLOAT outputs, vocoder plugin only
supports S16 at 44100. Indeed some code for this came from the vocoder plugin.
The converter type should be made into a config option. A perfectionist
would also ensure that if the plugin is disabled mid-stream, it cuts off at a
zero crossing, it currently does not.