Jack output volume change with zero crossing respect
This change makes its so that volume changes in the jack output plugin
only occur when the signal has zero energy. This prevents the 'zipper'
artifact from occuring when fading the volume up or down.

A lock has been inserted in the volume setting function as a kludge
to prevent any reentrant variable corruption. This is a kludge because
xmms2 core should handle the lock to prevent multiple clients from
changing the volume at exactly the same time.