OTHER: Jack output multichannel support DEVELOPER ONLY
This patch adds multichannel support to jack output, which had been
defined as two channel only prior. The default channel count is still
two, but with this patch you can set jack.channels to 1,2,3,4,5 or 6.
Which are mapped / set in accordance to what I am seeing as the defaults
in juhovh's tree.

This is just a preliminary patch in preperation / cooperation with juhovh's
work adding better multichannel support and conversion options. It doesn't
make use of the channel mask, it simply sets its availible output mode to the
channel count configured. AS-IS it wont even work if the input file matches up
due to the wackness of the sampler converter in this branch. (Wont convert S32
to FLOAT if channels != 2, or something)

The volume setting / port registration code parts added here are quite verbose, but
once we know exactly what we want we can trim them down to size.

Another limitation is the channel config option is not watched, and thus the ports are
set up only when the output plugin is loaded.

In my opinion the config option could be expanded so that multiple values (including
word equivlients such as quad for 4 and 5.1 for 6 could be put in at one time. (Comma
seperated) If more than one value is present, that means the jack plugin could then allow
that format to be availible (via format_set), but instead of destroying the unused ports,
all of the ports stay around (whatever the highest count is, that is what is registered
with jack on plugin load), and are left connected, but for example only left/right will be
outputing sound if a format_set comes through, and the channel count is 6. In my opinion
is is wrong behavior for the jack output plugin to change the actual ports around /
disconnect them / etc unless the user explicity changes the channel count config option.

(note format_set method is currently not present)

(fyi it will not be a problem to change the port config while the plugin is running)

It is my opinion that the channel mask and mappings should account for channel counts
1 to 8. Eight channel material is out there on the interweb.

But in addition to that I think we should allow something like 128 potential channels,
even though anything above eight is clearly experimental. The most the flac format supports
is eight. One could consider the corner case of a musuem or art gallery peice that used
some kind of 128 channel upconverter plugin for interesting things. Believe it or not I
don't actually have 128 speakers.